+++ src/main/java/com/bungaebowling/server/post/repository/PostRepository.java
@@ -12,23 +12,24 @@ import java.util.List;
 @Repository
 public interface PostRepository extends JpaRepository<Post, Long> {
 
+    // 모집글들 조회
     @Query("SELECT p FROM Post p WHERE p.district.id =:districtId ORDER BY p.id DESC")
-    List<Post> findAllByDistrictIdOrderByIdDesc(@Param("districtId") Integer districtId, Pageable pageable);
+    List<Post> findAllByDistrictIdOrderByIdDesc(@Param("districtId") Long districtId, Pageable pageable);
 
     @Query("SELECT p FROM Post p WHERE p.district.id =:districtId AND p.id < :key ORDER BY p.id DESC")
-    List<Post> findAllByDistrictIdAndIdLessThanOrderByIdDesc(@Param("districtId") Integer districtId, @Param("key") Long key, Pageable pageable);
+    List<Post> findAllByDistrictIdAndIdLessThanOrderByIdDesc(@Param("districtId") Long districtId, @Param("key") Long key, Pageable pageable);
 
     @Query("SELECT p FROM Post p WHERE p.district.country.id =:countryId ORDER BY p.id DESC")
-    List<Post> findAllByCountryIdOrderByIdDesc(@Param("countryId") Integer countryId, Pageable pageable);
+    List<Post> findAllByCountryIdOrderByIdDesc(@Param("countryId") Long countryId, Pageable pageable);
 
     @Query("SELECT p FROM Post p WHERE p.district.country.id =:countryId AND p.id < :key ORDER BY p.id DESC")
-    List<Post> findAllByCountryIdAndIdLessThanOrderByIdDesc(@Param("countryId") Integer countryId, @Param("key") Long key, Pageable pageable);
+    List<Post> findAllByCountryIdAndIdLessThanOrderByIdDesc(@Param("countryId") Long countryId, @Param("key") Long key, Pageable pageable);
 
     @Query("SELECT p FROM Post p WHERE p.district.country.city.id =:cityId ORDER BY p.id DESC")
-    List<Post> findAllByCityIdOrderByIdDesc(@Param("cityId") Integer cityId, Pageable pageable);
+    List<Post> findAllByCityIdOrderByIdDesc(@Param("cityId") Long cityId, Pageable pageable);
 
     @Query("SELECT p FROM Post p WHERE p.district.country.city.id =:cityId AND p.id < :key ORDER BY p.id DESC")
-    List<Post> findAllByCityIdAndIdLessThanOrderByIdDesc(@Param("cityId") Integer cityId, @Param("key") Long key, Pageable pageable);
+    List<Post> findAllByCityIdAndIdLessThanOrderByIdDesc(@Param("cityId") Long cityId, @Param("key") Long key, Pageable pageable);
 
     @Query("SELECT p FROM Post p ORDER BY p.id DESC")
     List<Post> findAllOrderByIdDesc(Pageable pageable);
@@ -38,22 +39,22 @@ public interface PostRepository extends JpaRepository<Post, Long> {
 
     // 마감안된 모집글만 조회
     @Query("SELECT p FROM Post p WHERE p.district.id =:districtId AND p.isClose = FALSE ORDER BY p.id DESC")
-    List<Post> findAllByDistrictIdWithCloseFalseOrderByIdDesc(@Param("districtId") Integer districtId, Pageable pageable);
+    List<Post> findAllByDistrictIdWithCloseFalseOrderByIdDesc(@Param("districtId") Long districtId, Pageable pageable);
 
     @Query("SELECT p FROM Post p WHERE p.district.id =:districtId AND p.isClose = FALSE AND p.id < :key ORDER BY p.id DESC")
-    List<Post> findAllByDistrictIdAndIdLessThanWithCloseFalseOrderByIdDesc(@Param("districtId") Integer districtId, @Param("key") Long key, Pageable pageable);
+    List<Post> findAllByDistrictIdAndIdLessThanWithCloseFalseOrderByIdDesc(@Param("districtId") Long districtId, @Param("key") Long key, Pageable pageable);
 
     @Query("SELECT p FROM Post p WHERE p.district.country.id =:countryId AND p.isClose = FALSE ORDER BY p.id DESC")
-    List<Post> findAllByCountryIdWithCloseFalseOrderByIdDesc(@Param("countryId") Integer countryId, Pageable pageable);
+    List<Post> findAllByCountryIdWithCloseFalseOrderByIdDesc(@Param("countryId") Long countryId, Pageable pageable);
 
     @Query("SELECT p FROM Post p WHERE p.district.country.id =:countryId AND p.isClose = FALSE AND p.id < :key ORDER BY p.id DESC")
-    List<Post> findAllByCountryIdAndIdLessThanWithCloseFalseOrderByIdDesc(@Param("countryId") Integer countryId, @Param("key") Long key, Pageable pageable);
+    List<Post> findAllByCountryIdAndIdLessThanWithCloseFalseOrderByIdDesc(@Param("countryId") Long countryId, @Param("key") Long key, Pageable pageable);
 
     @Query("SELECT p FROM Post p WHERE p.district.country.city.id =:cityId AND p.isClose = FALSE ORDER BY p.id DESC")
-    List<Post> findAllByCityIdWithCloseFalseOrderByIdDesc(@Param("cityId") Integer cityId, Pageable pageable);
+    List<Post> findAllByCityIdWithCloseFalseOrderByIdDesc(@Param("cityId") Long cityId, Pageable pageable);
 
     @Query("SELECT p FROM Post p WHERE p.district.country.city.id =:cityId AND p.isClose = FALSE AND p.id < :key ORDER BY p.id DESC")
-    List<Post> findAllByCityIdAndIdLessThanWithCloseFalseOrderByIdDesc(@Param("cityId") Integer cityId, @Param("key") Long key, Pageable pageable);
+    List<Post> findAllByCityIdAndIdLessThanWithCloseFalseOrderByIdDesc(@Param("cityId") Long cityId, @Param("key") Long key, Pageable pageable);
 
     @Query("SELECT p FROM Post p WHERE p.isClose = FALSE ORDER BY p.id DESC")
     List<Post> findAllWithCloseFalseOrderByIdDesc(Pageable pageable);
